fix(pos): erreur « could not determine data type of parameter $5 » bloquante au 1er encaissement
Cause racine identifiée par test E2E backend : la route /api/sales/quick
appelle posSessionService.ensureForSale qui fait un INSERT INTO pos_sessions
avec plusieurs params NULL non castés (gps_lat/lng/accuracy quand le
GPS est refusé ou indisponible, device_id et notes par défaut). pg ne
peut alors pas inférer le type, et l'erreur remonte sur le 5e param
(gps_lat) — message trompeur car les vrais coupables sont la chaîne
d'inférence cassée par les premiers nulls.

Le bug se manifestait pile au tout premier checkout d'un vendeur sur
un stand : la session POS n'existait pas encore, ensureForSale tentait
l'INSERT, et l'écran d'encaissement affichait le pg error.

Fix : casts explicites sur tous les paramètres potentiellement NULL
dans 3 endroits :

  - pos-session-service.ts (INSERT INTO pos_sessions) : $1-2::uuid,
    $4::varchar, $5-7::numeric, $8::text, $9::uuid. Le CASE WHEN $5
    IS NOT NULL bénéficie aussi du cast pour l'inférence du timestamp.

  - api/sales/quick/route.ts (INSERT INTO sales) : $3::uuid (client_id
    nullable en vente anonyme), $10::text (notes), $11-15::uuid (FK
    obligatoires + loyalty_rule_id nullable).
    + INSERT INTO sale_payments : $2,$5-8::uuid, $9::text.
    + INSERT INTO sale_items : $2-3::uuid.
    La route a son propre code d'INSERT (distinct de SaleService.create)
    qui n'avait pas reçu les casts du commit précédent.

  - stock-service.ts (UPDATE stock_items via decreaseStockOutlet) :
    COALESCE($3::numeric, minimum_stock) et $4::numeric. Le décrément
    de stock après vente passait par ce chemin et plantait quand
    minimumStock/maximumStock arrivaient en undefined → null.

  - sale-service.ts (boucle items) : résolution dual-id du productId
    (UUID ou business code PROD-…) ajoutée avant l'INSERT sale_items
    et le decreaseStockOutlet. Le POS passe l'UUID donc l'usage actuel
    n'est pas affecté, mais le service est désormais robuste si appelé
    avec un slug.

Le user doit hard-refresh sa page après push (Safari iOS cache parfois
le bundle JS — Turbopack a remonté un avertissement « stale » lié à
l'AppTopBar du commit précédent).

Vérifs : tsc 0 erreur côté prod. Bug pos_sessions confirmé corrigé
par test SELECT.

+++ lib/modules/stock/stock-service.ts
@@ -145,11 +145,14 @@ export class StockService {
     const newUnitCost = input.unitCost !== undefined ? input.unitCost : existing.UnitCost;
     const newTotalValue = newQuantity * newUnitCost;
 
+    // Casts explicites : pg ne peut pas inférer le type d'un paramètre
+    // null non casté dans COALESCE — symptôme « could not determine
+    // data type of parameter ». minimum/maximum_stock sont DECIMAL(10,3).
     const r = await db.query(
       `UPDATE stock_items
        SET quantity = $2,
-           minimum_stock = COALESCE($3, minimum_stock),
-           maximum_stock = COALESCE($4, maximum_stock),
+           minimum_stock = COALESCE($3::numeric, minimum_stock),
+           maximum_stock = COALESCE($4::numeric, maximum_stock),
            unit_cost = $5,
            total_value = $6,
            updated_at = CURRENT_TIMESTAMP
